+++ src/lib/components/ui/SubtitleFramePreview.svelte
@@ -1,8 +1,9 @@
 <script lang="ts">
 import { perma_state } from "$lib/persistent-storage.svelte"
 import { temp_state } from "$lib/temporary-storage.svelte"
+import { available_fonts } from "$lib/utils/fonts"
+import { render_ass_frame_preview } from "$lib/utils/video-processing"
 // See http://www.tcax.org/docs/ass-specs.htm
-import { available_fonts, render_ass_frame_preview } from "$lib/utils/subtitle-processing"
 import FontPreview from "../FontPreview.svelte"
 
 let sample_text = $state("The quick brown fox jumps over the lazy dog")
